==> notebooks/00_Data_Reduction_Scripts.ipynb <==
# 00 - Data generation and reduction

## Data files needed to run this notebook:

- Directory with metal songs downloaded from  and place in: ./data/metal_lyric/...
- ./data/lyrics-data.csv

## Settings:
none

# Import general purpose packages
import numpy as np 
import pandas as pd 
import sys
import json
import os
import re
import random

## Walk the directory tree and get the information

def get_all_metal_songs ():
  result = []
  for dirname, _, filenames in os.walk('./data/metal_lyrics'):
    for filename in filenames:
      result.append(os.path.join(dirname, filename))
  return result

# Get the songs
metal_files = get_all_metal_songs()

# for (i,x) in enumerate(metal_files[0].split('/')):
#   print(i,x)
  
split =   metal_files[10030].split('/')
song_idx = len(split)-1
artist_idx = len(split)-3

# TODO: empty lyrics remove
# TODO: strip white spaces


## Create subset

# Generate Dataframe with

# select a subsample

k = 30000

random.seed(42)
metal_files_short = random.choices(metal_files, k=k)

def show_status(i, total, percentage=10):
  if (i%(total/(100/percentage))==0):
    print(f"{round(i/total*100)}%")

# this process is getting really slow after 35000 records

artist = []
song = []
lyric = []

batch_size = 1000
for j in range(round(k/batch_size)):
  artist_tmp = []
  song_tmp = []
  lyric_tmp = []
  print(j*batch_size, (j+1)*batch_size)
  for (i,path) in enumerate(metal_files_short[j*batch_size: (j+1)*batch_size]):
    
    # format of file name: <prefix directories>/<letter>/<artist>/<album>/<track #>.<track name>.txt 
#     show_status(i, k, 5)
    p = path.split('/')

    song_idx = len(p)-1
    artist_idx = len(p)-3

    s = p[song_idx].split('.')
    artist_tmp.append(p[artist_idx].title())
    song_tmp.append(s[1].strip().title())

    # read file content for lyric
    f = open (path, 'r')
    lyric_tmp.append(f.read())
    f.close()
  pd.DataFrame({'Genre': 'Metal', 'Artist' : artist_tmp, 'Song' : song_tmp,'Lyric' : lyric_tmp}).to_csv(f"./tmp/metal_batch_{j}.csv")
    
#   artist.append(artist_tmp)
#   song.append(song_tmp)
#   lyric.append(lyric_tmp)
    
# create a dataframe for the data
# metal_songs = pd.DataFrame({'Genre': 'Metal', 'Artist' : artist, 'Song' : song,'Lyric' : lyric})
# metal_songs.head()

metal_songs = pd.DataFrame(columns=["Genre", "Artist", "Song", "Lyric"])

for j in range(round(k/batch_size)):
  metal_songs = df_tmp.append(pd.read_csv(f"./tmp/metal_batch_{j}.csv"))
  
  
metal_songs = metal_songs.reset_index()

metal_songs = metal_songs.drop(["index", "Unnamed: 0"], axis=1) 
metal_songs

## Write Subset

filename="metal_songs_TEMP"
path = "./data/"

metal_songs.to_csv(f"{path}{filename}.csv")

import pickle
with open(f"{path}{filename}.pkl", "wb") as fp: 
  pickle.dump(metal_songs, fp)

# create subset for Github 

df_lyrics = pd.read_csv(f"{path}lyrics-data.csv")

df_lyrics.shape

k = 75000

random.seed(42)
df_lyrics_small = df_lyrics.sample(n=k, replace = False, random_state=42, axis=0)
df_lyrics_small

df_lyrics_small.to_csv(f"{path}lyrics-data-small.csv")
