==> notebooks/02_DataPrep.ipynb <==
# 02 Final Data Prep - part 1

The data processing consist of 3 steps:

1. initial data collection and cleaning
2. creation of the BERT tokens. The creation of the tokens can take a long time so run this on COLAB with GPU enabled.
3. combining the tokens with the cleaned dataset

## Data files needed to run this notebook:
- `metal_songs.csv`
- `artist-data.csv`
- `lyrics-data.csv`

## Settings:
- set `COLAB = True` if you run this on Colab. Data can be placed in the root directory

# setup
import sys
import subprocess
import pkg_resources
from collections import Counter
import re
from numpy import log, mean, matmul


required = {'spacy', 'scikit-learn', 'numpy', 
            'pandas', 'torch', 'matplotlib',
            'transformers', 'allennlp==0.9.0'}
installed = {pkg.key for pkg in pkg_resources.working_set}
missing = required - installed

if missing:
    python = sys.executable
    subprocess.check_call([python, '-m', 'pip', 'install', *missing], stdout=subprocess.DEVNULL)
import spacy
import numpy as np
import pandas as pd

# SciKit Learn
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import normalize
from sklearn.metrics import accuracy_score
from sklearn.svm import LinearSVC
from sklearn.svm import SVC


# Spacy
from spacy.lang.en import English
en = English()

# !python -m spacy download en_core_web_md # includes GloVe Vectors
# !python -m spacy download en_core_web_sm
# !python -m spacy download en

# import en_core_web_sm
# import en_core_web_md


# PyTorch
import torch
# import torch.nn as nn
from torch.utils.data import TensorDataset, DataLoader



# File managment
import os
from os import listdir
from pathlib import Path
import pickle
import gzip

LOAD_DATA = False # read save data or regenerate data
SAVE_DATA = False # overwrite generated data? 

COLAB = False

if COLAB:
  # Google Colab
  path = "./"
  device = torch.device("cuda:0") # use GPU, change 
else:
  # Laptop
  path = "./data/"
  device = torch.device("cpu")
#   !pip install ipywidgets
#   !jupyter nbextension enable --py widgetsnbextension


def save_pickle(filename, data):
    with open(f"{path}{filename}", "wb") as fp: 
      pickle.dump(data, fp)

def load_pickle(filename):
    with open(f"{path}{filename}", 'rb') as f:
      return pickle.load(f)
    
def save_file(filename, train_tokens, test_tokens, val_tokens):
  data = {"train_tokens": train_tokens, "test_tokens" : test_tokens, "val_tokens": val_tokens}

  with open(f"{path}{filename}", "wb") as fp: 
    pickle.dump(data, fp)
    
def load_file(filename):
  with open(f"{path}{filename}", 'rb') as f:
      all_text = pickle.load(f)
      return (all_text["train_tokens"], all_text["test_tokens"], all_text["val_tokens"])


# Read data
df_artist =  pd.read_csv(f'{path}artists-data.csv')

df_lyrics = pd.read_csv(f'{path}lyrics-data.csv')

df_metal = pd.read_csv(f'{path}metal_songs.csv')

## Lyrics

df_lyrics_en = df_lyrics.query("Idiom == 'ENGLISH'")
df_lyrics_en = df_lyrics_en.drop(["SLink", "Idiom"], axis=1)

# optional filter #maybe filter Rock or Pop
df_artist_sm = df_artist.query("Genre == 'Rock' | Genre == 'Hip Hop' | Genre == 'Pop'")


## Artist

df_artist_sm_red = df_artist_sm.drop(["Genres", "Popularity", "Songs"], axis = 1)
df_artist_sm_red

# drop duplicates
df_artist_dedup = df_artist_sm_red.drop(df_artist_sm_red[df_artist_sm_red.Link.duplicated()].index)
df_artist_sm_red

## Metal

df_metal = df_metal[df_metal["Lyric"]!= ' Instrumental   ']
df_metal

df_metal = df_metal[~df_metal["Lyric"].isnull()]

l = df_metal["Lyric"].apply(lambda x: x.strip())
l

df_metal = df_metal[l!='']

df_metal2 = pd.DataFrame({"SName" : df_metal["Song"],	
                          "Lyric": df_metal["Lyric"],  	
                          "Artist" : df_metal["Artist"],  	
                          "Genre": df_metal["Genre"]})
                         
                       
                         
df_metal2 

## Join

# join the two data frames
df_lyrics_prejoin = df_lyrics_en.set_index('ALink')
df_artist_prejoin = df_artist_dedup.set_index('Link')
df_total = df_lyrics_prejoin.join(df_artist_prejoin)

# drop na values that couldn't be matched
df_total = df_total[~df_total["Genre"].isnull()]
df_total

# add metal dataset
df_total = df_total.append(df_metal2, ignore_index=True)
df_total

## Additional Cleaning

import re

def clean_lyrics(df, field):
  y = [re.sub('\[.*?\]', '', t) for t in df[field]]
  z = [re.sub('\(.*?\)', '', t) for t in y]
  df[field] = z
  return df

df_total_cleaned = clean_lyrics(df_total, "Lyric")
df_total_cleaned.head()

df_total_cleaned.shape

df_total_cleaned["Genre"].value_counts()

file_name = 'df_total_cleaned'
df_total_cleaned.to_csv(f"{path}{file_name}.csv")

save_pickle(f"{file_name}.pkl", df_total_cleaned  )

pickle.dump(df_total_cleaned, gzip.open(f'{path}{file_name}.pkl.gz', 'wb'))
